files: move commands to ws folder
